Initial commit
Rock, Paper, Scissors, Lizard, Spock is a game of chance that expands. It is first
used to settle a dispute about what to watch on TV between Sheldon and Raj in "The Lizard-Spock Expansion". It is mentioned again in "The Rothman Disintegration", where Sheldon explains the rules to Penny and Barry Kripke.
The rules are:
● Scissors cuts Paper
● Paper covers Rock
● Rock crushes Lizard
● Lizard poisons Spock
● Spock smashes Scissors
● Scissors decapitates Lizard
● Lizard eats Paper
● Paper disproves Spock
● Spock vaporizes Rock
● Rock crushes Scissors